+++ axelor-business-project/src/main/java/com/axelor/apps/businessproject/service/InvoicingProjectService.java
@@ -40,6 +40,7 @@ import com.axelor.apps.businessproject.exception.IExceptionMessage;
 import com.axelor.apps.hr.db.ExpenseLine;
 import com.axelor.apps.hr.db.TimesheetLine;
 import com.axelor.apps.hr.db.repo.ExpenseLineRepository;
+import com.axelor.apps.hr.db.repo.ExpenseRepository;
 import com.axelor.apps.hr.db.repo.TimesheetLineRepository;
 import com.axelor.apps.hr.db.repo.TimesheetRepository;
 import com.axelor.apps.hr.service.expense.ExpenseService;
@@ -210,7 +211,8 @@ public class InvoicingProjectService {
             false,
             saleOrderLine,
             null,
-            null) {
+            null,
+            false) {
 
           @Override
           public List<InvoiceLine> creates() throws AxelorException {
@@ -259,7 +261,8 @@ public class InvoicingProjectService {
             false,
             null,
             purchaseOrderLine,
-            null) {
+            null,
+            false) {
           @Override
           public List<InvoiceLine> creates() throws AxelorException {
 
@@ -319,6 +322,7 @@ public class InvoicingProjectService {
             PriceListLineRepository.AMOUNT_TYPE_NONE,
             project.getPrice().multiply(project.getQty()),
             null,
+            false,
             false) {
 
           @Override
@@ -386,9 +390,12 @@ public class InvoicingProjectService {
                 Beans.get(SaleOrderLineRepository.class)
                     .all()
                     .filter(
-                        "self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND self.saleOrder.creationDate < ?2",
+                        "self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND self.saleOrder.creationDate < ?2"
+                            + " AND (self.saleOrder.statusSelect = ?3 OR self.saleOrder.statusSelect = ?4)",
                         project,
-                        invoicingProject.getDeadlineDate())
+                        invoicingProject.getDeadlineDate(),
+                        SaleOrderRepository.STATUS_ORDER_CONFIRMED,
+                        SaleOrderRepository.STATUS_ORDER_COMPLETED)
                     .fetch());
 
         invoicingProject
@@ -397,7 +404,8 @@ public class InvoicingProjectService {
                 Beans.get(PurchaseOrderLineRepository.class)
                     .all()
                     .filter(
-                        "self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND self.purchaseOrder.orderDate < ?2",
+                        "self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND self.purchaseOrder.orderDate < ?2"
+                            + " AND (self.purchaseOrder.statusSelect = 3 OR self.purchaseOrder.statusSelect = 4)",
                         project,
                         invoicingProject.getDeadlineDate())
                     .fetch());
@@ -408,7 +416,8 @@ public class InvoicingProjectService {
                 Beans.get(TimesheetLineRepository.class)
                     .all()
                     .filter(
-                        "self.timesheet.statusSelect = 3 AND self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND self.date < ?2",
+                        "self.timesheet.statusSelect = ?1 AND self.project = ?2 AND self.toInvoice = true AND self.invoiced = false AND self.date < ?3",
+                        TimesheetRepository.STATUS_VALIDATED,
                         project,
                         invoicingProject.getDeadlineDate())
                     .fetch());
@@ -419,9 +428,12 @@ public class InvoicingProjectService {
                 Beans.get(ExpenseLineRepository.class)
                     .all()
                     .filter(
-                        "self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND self.expenseDate < ?2",
+                        "self.project = ?1 AND self.toInvoice = true AND self.invoiced = false AND self.expenseDate < ?2"
+                            + " AND (self.expense.statusSelect = ?3 OR self.expense.statusSelect = ?4",
                         project,
-                        invoicingProject.getDeadlineDate())
+                        invoicingProject.getDeadlineDate(),
+                        ExpenseRepository.STATUS_VALIDATED,
+                        ExpenseRepository.STATUS_REIMBURSED)
                     .fetch());
 
         invoicingProject
@@ -441,8 +453,11 @@ public class InvoicingProjectService {
                 Beans.get(SaleOrderLineRepository.class)
                     .all()
                     .filter(
-                        "self.project = ?1 AND self.toInvoice = true AND self.invoiced = false",
-                        project)
+                        "self.project = ?1 AND self.toInvoice = true AND self.invoiced = false"
+                            + " AND (self.saleOrder.statusSelect = ?2 OR self.saleOrder.statusSelect = ?3)",
+                        project,
+                        SaleOrderRepository.STATUS_ORDER_CONFIRMED,
+                        SaleOrderRepository.STATUS_ORDER_COMPLETED)
                     .fetch());
 
         invoicingProject
@@ -451,7 +466,8 @@ public class InvoicingProjectService {
                 Beans.get(PurchaseOrderLineRepository.class)
                     .all()
                     .filter(
-                        "self.project = ?1 AND self.toInvoice = true AND self.invoiced = false",
+                        "self.project = ?1 AND self.toInvoice = true AND self.invoiced = false"
+                            + " AND (self.purchaseOrder.statusSelect = 3 OR self.purchaseOrder.statusSelect = 4)",
                         project)
                     .fetch());
 
@@ -461,7 +477,8 @@ public class InvoicingProjectService {
                 Beans.get(TimesheetLineRepository.class)
                     .all()
                     .filter(
-                        "self.timesheet.statusSelect = 3 AND self.project = ?1 AND self.toInvoice = true AND self.invoiced = false",
+                        "self.timesheet.statusSelect = ?1 AND self.project = ?2 AND self.toInvoice = true AND self.invoiced = false",
+                        TimesheetRepository.STATUS_VALIDATED,
                         project)
                     .fetch());
 
@@ -471,8 +488,11 @@ public class InvoicingProjectService {
                 Beans.get(ExpenseLineRepository.class)
                     .all()
                     .filter(
-                        "self.project = ?1 AND self.toInvoice = true AND self.invoiced = false",
-                        project)
+                        "self.project = ?1 AND self.toInvoice = true AND self.invoiced = false"
+                            + " AND (self.expense.statusSelect = ?2 OR self.expense.statusSelect = ?3)",
+                        project,
+                        ExpenseRepository.STATUS_VALIDATED,
+                        ExpenseRepository.STATUS_REIMBURSED)
                     .fetch());
 
         invoicingProject