+++ axelor-sale/src/main/java/com/axelor/apps/sale/web/SaleOrderController.java
@@ -577,4 +577,17 @@ public class SaleOrderController {
             .getPriceListDomain(saleOrder.getClientPartner(), PriceListRepository.TYPE_SALE);
     response.setAttr("priceList", "domain", domain);
   }
+
+  public void removeSubLines(ActionRequest request, ActionResponse response) {
+    try {
+      SaleOrder saleOrder = request.getContext().asType(SaleOrder.class);
+      response.setValue(
+          "saleOrderLineList",
+          Beans.get(SaleOrderComputeService.class)
+              .removeSubLines(saleOrder.getSaleOrderLineList()));
+    } catch (Exception e) {
+      TraceBackService.trace(response, e);
+      response.setReload(true);
+    }
+  }
 }