+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/SaleOrderStockServiceImpl.java
@@ -318,6 +318,10 @@ public class SaleOrderStockServiceImpl implements SaleOrderStockService {
         stockMoveLine.setReservedQty(saleOrderLine.getReservedQty());
       }
 
+      stockMoveLine.setLineTypeSelect(saleOrderLine.getTypeSelect());
+      stockMoveLine.setPackPriceSelect(saleOrderLine.getPackPriceSelect());
+      stockMoveLine.setIsSubLine(saleOrderLine.getIsSubLine());
+
       return stockMoveLine;
     } else if (saleOrderLine.getTypeSelect() == SaleOrderLineRepository.TYPE_PACK) {
       StockMoveLine stockMoveLine =