+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/invoice/InvoiceServiceSupplychainImpl.java
@@ -36,14 +36,17 @@ import com.axelor.db.Query;
 import com.axelor.exception.AxelorException;
 import com.axelor.inject.Beans;
 import com.google.inject.Inject;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-public class InvoiceServiceSupplychainImpl extends InvoiceServiceImpl {
+public class InvoiceServiceSupplychainImpl extends InvoiceServiceImpl
+    implements InvoiceServiceSupplychain {
 
   @Inject
   public InvoiceServiceSupplychainImpl(
@@ -133,5 +136,61 @@ public class InvoiceServiceSupplychainImpl extends InvoiceServiceImpl {
           .flatMap(move -> move.getMoveLineList().stream())
           .collect(Collectors.toList());
     }
+  }
+
+  @Override
+  public List<InvoiceLine> addSubLines(List<InvoiceLine> invoiceLine) {
+
+    if (invoiceLine == null) {
+      return invoiceLine;
+    }
+
+    List<InvoiceLine> lines = new ArrayList<InvoiceLine>();
+    lines.addAll(invoiceLine);
+    for (InvoiceLine line : lines) {
+      if (line.getSubLineList() == null) {
+        continue;
+      }
+      for (InvoiceLine subLine : line.getSubLineList()) {
+        if (subLine.getInvoice() == null) {
+          invoiceLine.add(subLine);
+        }
+      }
+    }
+    return invoiceLine;
+  }
+
+  @Override
+  public List<InvoiceLine> removeSubLines(List<InvoiceLine> invoiceLines) {
+
+    if (invoiceLines == null) {
+      return invoiceLines;
+    }
+
+    List<InvoiceLine> subLines = new ArrayList<InvoiceLine>();
+    for (InvoiceLine packLine : invoiceLines) {
+      if (packLine.getTypeSelect() == 2 && packLine.getSubLineList() != null) {
+        packLine.getSubLineList().removeIf(it -> it.getId() != null && !invoiceLines.contains(it));
+        packLine.setTotalPack(
+            packLine
+                .getSubLineList()
+                .stream()
+                .map(it -> it.getExTaxTotal())
+                .reduce(BigDecimal.ZERO, BigDecimal::add));
+        subLines.addAll(packLine.getSubLineList());
+      }
+    }
+    Iterator<InvoiceLine> lines = invoiceLines.iterator();
+
+    while (lines.hasNext()) {
+      InvoiceLine subLine = lines.next();
+      if (subLine.getId() != null
+          && subLine.getParentLine() != null
+          && !subLines.contains(subLine)) {
+        lines.remove();
+      }
+    }
+
+    return invoiceLines;
   }
 }